Renamed CopyAndReleaseBufferDataCommand and CopyAndReleaseImageDataCommand to CopyAndReleaseBuffer and CopyAndReleaseImage
