+++ src/vsg/CMakeLists.txt
@@ -41,8 +41,8 @@ set(SOURCES
     commands/BlitImage.cpp
     commands/CopyImage.cpp
     commands/CopyImageToBuffer.cpp
-    commands/CopyAndReleaseBufferDataCommand.cpp
-    commands/CopyAndReleaseImageDataCommand.cpp
+    commands/CopyAndReleaseBuffer.cpp
+    commands/CopyAndReleaseImage.cpp
     commands/PipelineBarrier.cpp
     commands/Event.cpp
     commands/NextSubPass.cpp
